Produces a working treemap to level 4
I added a root variable so that it's an option in the trail, and also fixed issue where including some portfolios resulted in no treemap rendering. Issue was IDs were duplicated, which I thought I had fixed by appending with a string, but there was further duplication even with that, so added portfolio name to ID in those situations, which solved it.